Add a type to represent asynchronous operations
The `Laodable<T>` sum type represents an asynchronous operation that may
take a varying amount of time.

It is composed of 4 variants, each having it's own constructor:
  - NotLoaded()
  - Loading()
  - Loaded(T)
  - Failed(Error)

It is impossible to unwrap a `Laodable<T>` directly, as all members are
private.  If one wishes to use the information contained within one of
its variants, patterns matching must be used:
```ts
declare const loadedUser: Loadable<User>

const userElement = loadedUser.match({
  Loaded (user: User) {
    return (
      <UserCard user={user} />
    )
  },
  Error (error: Error) {
    return (
      <ErrorMessage error={error} />
    )
  },
  _ () {
    return (
      <LoadingBar type='indefinite' />
    )
  }
})
```

In the future, a fifth variant `Reloading` might be added, expressing
the fact that we have a value that is also being reloaded.